requirements.txt
matplotlib==3.0.2
numpy==1.15.4
pandas==0.23.4
scikit-learn==0.20.1
PyQt==5.9.2